[ABW]: Chain objects can now use custom constructors - selected constructors (that were not typical) use that mechanism All typical chain object constructors replaced with CHAINBASE_DEFAULT_CONSTRUCTOR macro Chain object's id_type now only used for mandatory id; members of type <chainobject>_id_type are now id_type_ref Note: id_type has to be explicitly initialized with integer while id_type_ref can only be default or assigned from compatible id_type

+++ libraries/chain/include/hive/chain/util/sps_helper.hpp
@@ -82,7 +82,7 @@ class sps_helper
          /// Now remove all votes specific to given proposal.
          auto propI = byVoterIdx.lower_bound(boost::make_tuple(proposal->proposal_id, account_name_type()));
 
-         while(propI != byVoterIdx.end() && static_cast< size_t >( propI->proposal_id ) == static_cast< size_t >( proposal->proposal_id ) )
+         while(propI != byVoterIdx.end() && propI->proposal_id == proposal->proposal_id )
          {
             auto result_itr = checker< ByProposalType, true/*Loop*/ >( proposal, proposalIndex, obj_perf );
             if( obj_perf.done )
